+++ samples/Vulkan-Tutorial/main.cpp
@@ -983,7 +983,7 @@ public:
     void recordCommands() {
         vk::CommandBufferAllocateInfo allocInfo = {};
         allocInfo.commandPool = commandPool.get();
-        allocInfo.commandBufferCount = swapchain->images().size();
+        allocInfo.commandBufferCount = static_cast<uint32_t>(swapchain->images().size());
         allocInfo.level = vk::CommandBufferLevel::Primary;
 
         commandBuffers = commandPool->allocate(allocInfo);
